Add proper OG images
With better styling, component icons and a special variant for the landing page.

The new aproach now uses a Next.js page which is screenshotted, instead of rendering a component directly on the server. This is done, because otherwise we don't get correct styling (e.g. vanilla-extract styles don't load).