Sidebar: Fix erratic selected-item indicator
Store the active sidebar element in a separate store. With the previous approach, where the active item was determined in Item.ts, the scroll-handling from Next.js interfered with the measuring of framer-motion, leading to wrong layout animations of the active-item indicator.

With this aproach the item is updated lazily (in a useEffect) so it doesn't the layout measuring from framer-motion doesn't interfere with the scroll-handling of Next.js